--- FE/src/components/login/ui/StyledLoginButton.tsx
@@ -1,32 +1,0 @@
-import classNames from "classnames";
-import Image from "next/image";
-import Link from "next/link";
-
-const defaultStyle = "flex w-64 justify-center gap-4 rounded-3xl py-3";
-
-const colors = {
-  slack: "bg-slack text-white text-paragraph",
-  guest: "bg-primary text-black text-paragraph",
-};
-
-interface StyledLoginButtonProps {
-  linkUrl: string;
-  buttonText: string;
-  imageUrl: string;
-  color: keyof typeof colors;
-}
-
-export default function StyledLoginButton({
-  linkUrl,
-  buttonText,
-  imageUrl,
-  color,
-}: StyledLoginButtonProps) {
-  const buttonStyle = classNames(defaultStyle, colors[color]);
-  return (
-    <Link className={buttonStyle} href={linkUrl}>
-      <Image src={imageUrl} alt="슬랙 로고" width={24} height={24} />
-      <p className="text-center font-semibold">{buttonText}</p>
-    </Link>
-  );
-}
